Rename triggers.tsx to webhooks.tsx
